+++ server/src/docs/auth.docs.ts
