+++ server/src/routes/auth.routes.ts
