+++ server/src/routes/docs.routes.ts
